chore: 🤖 move test dir
