+++ packages/amplify-graphql-api-construct-tests/src/__tests__/relationships/composite-pk/graphql/schema-primary-cpk-1sk.graphql
