+++ packages/amplify-graphql-api-construct-tests/src/__tests__/relationships/composite-pk/graphql/schema-related-cpk-2sk.graphql
